Change location of files
