+++ Lee/src/main/java/com/example/Lee/domain/Answer.java
@@ -1,4 +1,4 @@
-package domain;
+package com.example.Lee.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -9,14 +9,17 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @Entity
-public class Question {
+public class Answer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    @Column(length = 200)
-    private String subject;
+
     @Column(columnDefinition = "TEXT")
     private String content;
 
     private LocalDateTime createDate;
+
+    @ManyToOne
+    private Question question;
+
 }
